Make non-admin error message less specific
The same warning is shown whether the script has been run from
PowerShell or cmd.  The script works fine in an administrative command
prompt, so no need to send users off to PowerShell if they weren't already
using it.  Additionally, starting an administrative version of either
environment can be performed several ways, many of which involve no
clicking at all, and most users should be familiar enough with running
programs as an Administrator.